feat(lst parsing): change the parsing methode to create a chain list with all the number\n\nHowever there is a seg fault when more 7 character, and only the first number wath add to the stacka.\nFor exemple with ./push_swap '5 6 7' the output was '5', need a fix

+++ Push-Swap/libft/conv_utils/ft_antoi.c
@@ -12,7 +12,7 @@
 
 # include "../libft.h"
 
-int ft_antoi(char *args, int *i)
+int ft_antoi(char *args, size_t *i)
 {
 	int	res;
 	int	sign;
