+++ Push-Swap/libft/lst_utils/ft_lstdelone.c
@@ -1,6 +1,6 @@
 #include "../libft.h"
 
-void	ft_lstdelone(t_list *lst, void (*del)(int))
+void	ft_lstdelone(t_list *lst, void (*del)(long))
 {
 	if (!lst || !del)
 		return ;
