+++ Push-Swap/libft/lst_utils/ft_lstiter.c
@@ -1,6 +1,6 @@
 #include "../libft.h"
 
-void	ft_lstiter(t_list *lst, void (*f)(int))
+void	ft_lstiter(t_list *lst, void (*f)(long))
 {
 	t_list	*tmp;
