+++ Push-Swap/libft/lst_utils/ft_lstmap.c
@@ -1,10 +1,10 @@
 #include "../libft.h"
 
-t_list	*ft_lstmap(t_list *lst, int (*f)(int), void (*del)(int))
+t_list	*ft_lstmap(t_list *lst, long (*f)(long), void (*del)(long))
 {
 	t_list	*new;
 	t_list	*tmp;
-	int		value;
+	long	value;
 
 	if (!lst || !f || !del)
 		return (NULL);
